+++ .github/workflows/basicrun.yml
@@ -1,0 +1,34 @@
+name: RunBasic
+
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    # The branches below must be a subset of the branches above
+    branches: [ master ]
+
+
+jobs:
+  checks:
+    name: basic_assess
+    runs-on: ubuntu-latest
+    steps:
+    - name: Checkout
+      uses: actions/checkout@master
+     
+    - name: Setup Go
+      uses: actions/setup-go@v2
+      with:
+        go-version: '1.17.6'
+
+    - name: Run build
+      run: go build github.com/${GITHUB_REPOSITORY}
+
+    - name: Run test
+      run: go test github.com/${GITHUB_REPOSITORY} -coverprofile=profile.cov
+    
+    - name: Coveralls
+      uses: shogo82148/actions-goveralls@v1
+      with:
+        path-to-profile: profile.cov
+        ignore: recordsorganiser.go
